tests/wpt/tests/html/canvas/offscreen/path-objects/2d.path.open.arc.html

<!DOCTYPE html>
<!-- DO NOT EDIT! This test has been generated by /html/canvas/tools/gentest.py. -->
<meta charset="UTF-8">
<title>OffscreenCanvas test: 2d.path.open.arc</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/html/canvas/resources/canvas-tests.js"></script>

<h1>2d.path.open.arc</h1>
<p class="desc">line caps should be drawn on open arcs</p>


<script>
var t = async_test("line caps should be drawn on open arcs");
var t_pass = t.done.bind(t);
var t_fail = t.step_func(function(reason) {
    throw reason;
});
t.step(function() {

  var canvas = new OffscreenCanvas(300, 300);
  var ctx = canvas.getContext('2d');

  ctx.beginPath();
  ctx.strokeStyle = 'red';
  ctx.lineWidth = 20;
  ctx.lineCap = 'square';
  ctx.arc(150, 150, 50, 0, 2 * Math.PI);
  ctx.stroke();
  // The pixel (209, 148) is part of the line cap, but not part of the circle.
  // Linecap is drawn with 360 degree arc, so the pixel is red.
  _assertPixelApprox(canvas, 209,148, 255,0,0,255, 1);
  t.done();

});
</script>
